requirements.txt
fastapi
uvicorn
python-multipart

aiogram
httpx

python-docx

numpy
pandas
openpyxl
scikit-learn

sentence-transformers
transformers
torch